Modify e_bank.h
This commit was made to enhance the user experience in the program. In
the earlier state of the program, the program immediately exits when a
user enters a wrong character-length, username or password when
inputting data. This was addressed by allowing the user another chance
to enter the correct data-length, username or password and then continue
with the program instead of the immediate exit from the program that
happens upon wrong data-length, username or password input.